chore: Refractored ``obj_to_dict``
